Remove region selection - NTC region is based on the admin's region - Refactor code - Other bug fixes
