+++ README.md
@@ -2,13 +2,11 @@
 
 ## Run via browser (for development)
 1. Open a terminal and make sure you are in the project folder (i.e. /var/git/netmesh-rfc6349-client)
-2. Run `python3 netmesh_rfc6349_app.py`
+2. Run `python3 run.py`
 3. Go to the link in the output of the above command (usually http://127.0.0.1:5000/)
 
 ## Run via desktop app (for release/production)
 ### Building the app
 1. Open a terminal and make sure you are in the project folder (i.e. /var/git/netmesh-rfc6349-client)
 2. Run `python3 netmesh_install.py`
-### Running the app
-1. Open `dist` folder (this is in the project folder)
-2. Open the '.desktop' file (or the file with a black icon)
+...
