+++ netmesh_rfc6349_app/main/utils/netmesh_config_file.py
@@ -56,7 +56,7 @@ class NetMeshConfigFile:
                     self._SECTION_NAME, "device_name")
             except Exception as ex:
                 print(ex)
-
+                
             return device_name
 
     class _NetMeshUsersConfig:
