+++ netmesh_rfc6349_app/test_measurement/routes.py
@@ -18,7 +18,6 @@ from netmesh_rfc6349_app.test_measurement.utils import run_process_script, get_n
 
 from netmesh_rfc6349_app.main.utils.netmesh_installer import get_app_current_version, app_resource_path
 from netmesh_rfc6349_app.main.utils import laptop_info
-from netmesh_rfc6349_app.device_registration.utils import get_device_info
 
 import netmesh_rfc6349_app.main.utils.log_settings as log_settings
 
@@ -893,145 +892,6 @@ def send_error():
         return Response(json.dumps({
             "error": error
         }), 500)
-
-
-@test_measurement.route('/get-admin-credentials-template', methods=['POST'])
-def get_admin_credentials_template():
-    return render_template('device_admin_credentials.html')
-
-
-@test_measurement.route('/log-admin', methods=['POST'])
-def log_admin():
-    # device_name = request.form.get('device-name')
-    admin_email = request.form.get('admin-email')
-    admin_password = request.form.get('admin-password')
-
-    print(admin_email, admin_password)
-
-    r = None
-    try:
-        r = requests.post(
-            url=f"{current_app.config['RESULTS_SERVER_API_URL']}/user/token/",
-            json={
-                "email": admin_email,
-                "password": admin_password,
-                "client": "admin"
-            }
-        )
-        r.raise_for_status()
-
-        token = r.json()['token']
-        session['admin-token'] = token
-    except requests.exceptions.RequestException as re:
-        error = "Unexpected error"
-        try:
-            error_json = json.loads(r.content)
-            if "non_field_errors" in error_json:
-                error = error_json["non_field_errors"]
-            else:
-                error = r.content
-        except Exception as ex:
-            print(ex)
-            error = str(re)
-
-        return render_template('device_admin_credentials.html', error=error), 400
-
-    return render_template('device_region.html', regions=netmesh_location.get_philippine_regions())
-
-
-@test_measurement.route('/get-device-details-template', methods=['POST'])
-def get_device_details_template():
-    region = request.form.get('region')
-    region_name = request.form.get('regionName')
-
-    print("ntc_region", region)
-    print("token", session['admin-token'])
-
-    regions = netmesh_location.get_philippine_regions()
-    try:
-        r = requests.get(
-            url=f"{current_app.config['RESULTS_SERVER_API_URL']}/user/",
-            params={
-                "ntc_region": region
-            },
-            headers={'Authorization': f"Token { session['admin-token'] }"}
-        )
-        r.raise_for_status()
-
-        print("users")
-        print(r.text)
-    except requests.exceptions.RequestException as re:
-        error = "Unexpected error"
-        try:
-            error_json = json.loads(r.content)
-            if "non_field_errors" in error_json:
-                error = error_json["non_field_errors"]
-            else:
-                error = r.content
-        except Exception as ex:
-            print(ex)
-            error = str(re)
-
-        return render_template('device_region.html', regions=regions, error=error), 400
-
-    return render_template('device_details_form.html', users=r.json(), region_name=region_name)
-
-
-@test_measurement.route('/device-confirmation-template', methods=['POST'])
-def device_confirmation_template():
-    region_name = request.form.get('regionName')
-    laptop_owner = request.form.get('laptopOwner')
-    laptop_owner_email = request.form.get('laptopOwnerEmail')
-    laptop_name = request.form.get('laptopName')
-
-    return render_template('device_confirmation.html',
-                           region_name=region_name,
-                           laptop_owner=laptop_owner,
-                           laptop_name=laptop_name,
-                           laptop_owner_email=laptop_owner_email,
-                           device_info=get_device_info())
-
-
-@test_measurement.route('/device-details', methods=['GET'])
-def device_details():
-    return render_template('device_details_form.html', users=[
-        {
-            "ntc_region": "3",
-            "id": 1_000_001,
-            "email": "romano.bourgeois@example.com",
-            "first_name": "Romano",
-            "last_name": "Bourgeois"
-        },
-        {
-            "ntc_region": "3",
-            "id": 1_000_002,
-            "email": "peyton.vasquez@example.com",
-            "first_name": "Peyton",
-            "last_name": "Vasquez"
-        },
-        {
-            "ntc_region": "3",
-            "id": 1_000_003,
-            "email": "gavin.cole@example.com",
-            "first_name": "Gavin",
-            "last_name": "Cole"
-        },
-        {
-            "ntc_region": "3",
-            "id": 1_000_004,
-            "email": "audrey.tucker@example.com",
-            "first_name": "Audrey",
-            "last_name": "Tucker"
-        },
-        {
-            "ntc_region": "3",
-            "id": 1_000_005,
-            "email": "jill.pearson@example.com",
-            "first_name": "Jill",
-            "last_name": "Pearson"
-        }
-    ])
-
 
 @test_measurement.route('/open-downloads-folder', methods=['POST'])
 def open_downloads_folder():
